+++ builders/drupal-postgres.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const _ = require('lodash');
-const LandoPostgres = require('./../node_modules/@lando/postgres/builders/postgres.js');
+const LandoPostgres = require('@lando/postgres/builders/postgres.js');
 
 // Builder
 module.exports = {